pyproject.toml
[tool.poetry]
name = "pytorch-cyclegan-and-pix2pix"
version = "0.1.0"
description = ""
authors = ["joshuachadney <joshua.chadney@satellitevu.com>"]
readme = "README.md"
packages = [{include = "pytorch_cyclegan_and_pix2pix"}]

[tool.poetry.dependencies]
python = "^3.10"
dominate = "2.6.0"
torch = "^2.2.2"
torchvision = "^0.17.2"
pillow = "^10.3.0"
visdom = "^0.2.4"
wandb = "^0.16.6"
numpy = "^1.26.4"
scipy = "^1.13.0"
awslambdaric = "^2.0.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
